refactor: 알콜, 스낵 패키지 분리
