+++ src/main/java/com/soolsul/soolsulserver/menu/alcohol/domain/AlcoholMenu.java
@@ -1,4 +1,4 @@
-package com.soolsul.soolsulserver.menu.domain;
+package com.soolsul.soolsulserver.menu.alcohol.domain;
 
 import com.soolsul.soolsulserver.common.domain.BaseTimeEntity;
 import lombok.AccessLevel;
@@ -8,9 +8,12 @@ import javax.persistence.Entity;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SnackMenu extends BaseTimeEntity {
+public class AlcoholMenu extends BaseTimeEntity {
 
-    private Long cost;
-    private String snackId;
+    private double cost;
+
+    private String alcoholId;
+
     private String barId;
+
 }
