+++ src/main/java/com/soolsul/soolsulserver/menu/snack/domain/Snack.java
@@ -1,4 +1,4 @@
-package com.soolsul.soolsulserver.menu.domain;
+package com.soolsul.soolsulserver.menu.snack.domain;
 
 import com.soolsul.soolsulserver.common.domain.BaseTimeEntity;
 import lombok.AccessLevel;
@@ -8,11 +8,8 @@ import javax.persistence.Entity;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Alcohol extends BaseTimeEntity {
+public class Snack extends BaseTimeEntity {
 
     private String name;
-
-    private int alcoholPercent;
-
-    private Long alcoholCategoryId;
+    private Long snackCategoryId;
 }
